TA-761 serverless katana

+++ scripts/build.sh
@@ -1,5 +1,17 @@
 #!/bin/bash -e
-rm -rf public/build
+
+echo "Building Server Code"
 rm -rf dist
 NODE_ENV=production $(npm bin)/babel src -d dist --ignore client/* --no-babelrc --env='{ env: { production: { targets: { node: "current" , "uglify" : "true"} } } }'  --plugins transform-object-rest-spread,transform-es2015-modules-commonjs
-NODE_ENV=production $(npm bin)/webpack --config webpack.config.production.js
+echo "Server Build Complete"
+
+if [ "$1" == "all" ]
+then 
+    echo "Building Webpack"
+    rm -rf public/build
+    NODE_ENV=production $(npm bin)/webpack --config webpack.config.production.js
+    echo "Webpack Build Complete"
+fi
+
+
+
